requirements.txt
flask
groq
openai
requests
Pillow
python-dotenv